vcl consitent use of long for corrdinate
most of length in vcl are calculated in 'long'
but array of X position tend to be in sal_Int32.
As a prep work to be able to support 'double'
as the base type of Device Coordinate, harmonize
the use of 'long' for non-float coordinate.

+++ vcl/win/source/gdi/winlayout.cxx
@@ -794,7 +794,7 @@ void SimpleWinLayout::ApplyDXArray( const ImplLayoutArgs& rArgs )
         if( j >= 0 )
         {
             nOldWidth += mpGlyphAdvances[ j ];
-            int nDiff = nOldWidth - pDXArray[ i ];
+            long nDiff = nOldWidth - pDXArray[ i ];
 
             // disabled because of #104768#
             // works great for static text, but problems when typing
